# openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/readiness": {
            "get": {
                "description": "Always public. Returns 200 only when the API and all dependencies required to serve traffic are available.",
                "operationId": "HealthController_readiness",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "API is ready for traffic.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "status": "ok",
                                        "timestamp": "2026-01-01T00:00:00.000Z",
                                        "uptimeSeconds": 120,
                                        "checks": {
                                            "application": {
                                                "status": "up"
                                            },
                                            "database": {
                                                "status": "up",
                                                "responseTimeMs": 3
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "A required dependency is unavailable."
                    }
                },
                "summary": "Traffic readiness probe",
                "tags": ["Health"]
            }
        },
...
    },
    "info": {
        "title": "Expense Splitter API",
        "description": "REST API for the Expense Splitter application",
        "version": "0.0.1",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "access-token": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
...
    }
}